+++ packages/putting-pals-core/src/leaderboard/leaderboard-service.ts
@@ -1,41 +1,29 @@
 import { LeaderboardClient } from "@putting-pals/pga-tour-api/leaderboard";
 import type { TourCode } from "@putting-pals/putting-pals-schema/types";
 import { CompetitionService } from "../competition/competition-service";
-import { ScheduleService } from "../schedule/schedule-service";
+import { TournamentResolver } from "../tournament/tournament-resolver";
 import { assertNever } from "../utils/type-utils";
 import { aggregateLeaderboard } from "./leaderboard-aggregator";
 import { transformLeaderboard } from "./leaderboard-transformer";
 
 export class LeaderboardService {
-  getLeaderboard(tourCode: TourCode, id?: string) {
+  async getLeaderboard(tourCode: TourCode, id?: string) {
+    const tournamentId = await this.resolveTournamentId(tourCode, id);
     switch (tourCode) {
       case "P":
-        return this.getPuttingPalsLeaderboard(id);
+        return this.getPuttingPalsLeaderboardById(tournamentId);
       case "R":
-        return this.getPgaTourLeaderboard(id);
+        return this.getPgaTourLeaderboardById(tournamentId);
       default:
         assertNever(tourCode);
     }
   }
 
-  private async getPuttingPalsLeaderboard(id?: string) {
-    if (id) {
-      return this.getPuttingPalsLeaderboardById(id);
-    } else {
-      const currentTournamentId =
-        await new ScheduleService().getCurrentTournamentId("P");
-      return this.getPuttingPalsLeaderboardById(currentTournamentId);
+  private async resolveTournamentId(tourCode: TourCode, id?: string) {
+    if (id === undefined) {
+      return await new TournamentResolver().getCurrentTournamentId(tourCode);
     }
-  }
-
-  private async getPgaTourLeaderboard(id?: string) {
-    if (id) {
-      return this.getPgaTourLeaderboardById(id);
-    } else {
-      const currentTournamentId =
-        await new ScheduleService().getCurrentTournamentId("R");
-      return this.getPgaTourLeaderboardById(currentTournamentId);
-    }
+    return id;
   }
 
   private async getPuttingPalsLeaderboardById(id: string) {
