+++ packages/putting-pals-core/src/schedule/schedule-service.ts
@@ -1,45 +1,14 @@
 import { ScheduleClient } from "@putting-pals/pga-tour-api/schedule";
-import { PgaTourWebScraper } from "@putting-pals/pga-tour-scaper/scraper";
 import type { TourCode } from "@putting-pals/putting-pals-schema/types";
 import { parseISO } from "date-fns";
 import { CompetitionService } from "../competition/competition-service";
 import { TournamentService } from "../tournament/tournament-service";
 import { NotFoundError } from "../utils/service-error";
 import { assertNever } from "../utils/type-utils";
-import { groupScheduleByYear } from "./schedule-grouper";
 import { transformSchedule } from "./schedule-transformer";
-import { transformScheduleYears } from "./schedule-years-transformer";
+import { ScheduleYearsService } from "./schedule-years-service";
 
 export class ScheduleService {
-  getScheduleYears(tourCode: TourCode) {
-    switch (tourCode) {
-      case "P":
-        return this.getPuttingPalsScheduleYears();
-      case "R":
-        return this.getPgaTourScheduleYears();
-      default:
-        assertNever(tourCode);
-    }
-  }
-
-  private async getPuttingPalsScheduleYears() {
-    const [pgaTourScheduleYears, puttingPalsHistoricalSchedule] =
-      await Promise.all([
-        this.getPgaTourScheduleYears(),
-        this.getPuttingPalsHistoricalSchedule(),
-      ]);
-    return pgaTourScheduleYears.filter((year) =>
-      puttingPalsHistoricalSchedule.some(
-        (schedule) => schedule.year === year.queryValue,
-      ),
-    );
-  }
-
-  private async getPgaTourScheduleYears() {
-    const scheduleYears = await new ScheduleClient().getScheduleYears();
-    return transformScheduleYears(scheduleYears);
-  }
-
   getSchedule(tourCode: TourCode, year?: string) {
     switch (tourCode) {
       case "P":
@@ -51,33 +20,13 @@ export class ScheduleService {
     }
   }
 
-  getCurrentTournamentId(tourCode: TourCode) {
-    switch (tourCode) {
-      case "P":
-        return this.getCurrentPuttingPalsTournamentId();
-      case "R":
-        return this.getCurrentPgaTourTournamentId();
-      default:
-        assertNever(tourCode);
-    }
-  }
-
-  private async getPuttingPalsHistoricalSchedule() {
-    const competitions = new CompetitionService().getCompetitions();
-    const tournamentIds = competitions.map(
-      (competition) => competition.tournamentId,
-    );
-    const tournaments = await new TournamentService().getTournaments(
-      tournamentIds,
-    );
-    return groupScheduleByYear(tournaments);
-  }
-
   private async getPuttingPalsSchedule(year?: string) {
     if (year) {
       return this.getPuttingPalsScheduleByYear(year);
     } else {
-      const scheduleYears = await this.getPuttingPalsScheduleYears();
+      const scheduleYears = await new ScheduleYearsService().getScheduleYears(
+        "P",
+      );
       const currentYear = scheduleYears.find((year) => year.current);
       if (currentYear === undefined) {
         throw new NotFoundError("No current year found");
@@ -131,23 +80,15 @@ export class ScheduleService {
           ),
         })),
     };
-
-    // return this.getPuttingPalsHistoricalSchedule().then((groupedSchedule) => {
-    //   const seasonSchedule = groupedSchedule.find(
-    //     (season) => season.year === year,
-    //   );
-    //   if (seasonSchedule === undefined) {
-    //     throw new NotFoundError(`Schedule for year=${year} not found`);
-    //   }
-    //   return seasonSchedule;
-    // });
   }
 
   private async getPgaTourSchedule(year?: string) {
     if (year) {
       return this.getPgaTourScheduleByYear(year);
     } else {
-      const scheduleYears = await this.getPgaTourScheduleYears();
+      const scheduleYears = await new ScheduleYearsService().getScheduleYears(
+        "R",
+      );
       const currentYear = scheduleYears.find((year) => year.current);
       if (currentYear === undefined) {
         throw new NotFoundError("No current year found");
@@ -159,51 +100,5 @@ export class ScheduleService {
   private async getPgaTourScheduleByYear(year: string) {
     const schedule = await new ScheduleClient().getSchedule(year);
     return transformSchedule(schedule);
-  }
-
-  private getCurrentPgaTourTournamentId() {
-    return new PgaTourWebScraper().getCurrentTournamentId();
-  }
-
-  private async getCurrentPuttingPalsTournamentId() {
-    const competitions = new CompetitionService().getCompetitions();
-
-    const oddsAvailableCompetition = competitions.find(
-      (competition) =>
-        competition.competitors.length === 0 &&
-        competition.paddyPowerId !== undefined,
-    );
-
-    if (oddsAvailableCompetition !== undefined) {
-      return oddsAvailableCompetition.tournamentId;
-    }
-
-    const inProgressCompetition = competitions.find(
-      (competition) =>
-        competition.competitors.length > 0 &&
-        competition.winnerId === undefined &&
-        competition.runnerUpId === undefined,
-    );
-
-    if (inProgressCompetition !== undefined) {
-      return inProgressCompetition.tournamentId;
-    }
-
-    const competitionIds = competitions
-      .filter((competition) => competition.competitors.length > 0)
-      .map((competition) => competition.tournamentId);
-
-    const tournaments = await new TournamentService().getTournaments(
-      competitionIds,
-    );
-    const currentTournament = tournaments.sort((a, b) =>
-      b.startDate.localeCompare(a.startDate),
-    )[0];
-
-    if (currentTournament === undefined) {
-      throw new NotFoundError("No tournament found");
-    }
-
-    return currentTournament.id;
   }
 }
